requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
cryptojwt==1.5.2
filelock==3.3.0
idna==3.2
oidcmsg==1.3.3.post1
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
PyYAML==5.4.1
readerwriterlock==1.0.9
requests==2.26.0
responses==0.14.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
